requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10
easyocr>=1.7.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0